Move prescriptions comparison logic to verification module
