+++ coordinator/tests/services/verification/compare-prescriptions.spec.ts
@@ -2,9 +2,8 @@
 
 import {readFileSync} from "fs"
 import path from "path"
-import {buildPrescription} from "../../../../models/common"
 import * as LosslessJson from "lossless-json"
-import {comparePrescriptions} from "../../../src/routes/dispense/verify-signature"
+import {comparePrescriptions} from "../../../src/services/verification/signature-verification"
 import {common, fhir} from "@models"
 
 const basePath = "../../../../examples/primary-care/repeat-dispensing/nominated-pharmacy/medical-prescriber/author/gmc/responsible-party/medication-list/din"
@@ -23,7 +22,7 @@ test("build prescription returns correct values", () => {
     "utf-8"
   )
   const sendFhir: fhir.Bundle = LosslessJson.parse(sendFhirStr)
-  const result = buildPrescription(sendFhir)
+  const result = common.buildPrescription(sendFhir)
   expect(result).toStrictEqual(defaultPrescription)
 })
